+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+
+## 1.0.1
+
+- Added a simple autocomplete feature for properties.
+![Autocomplete][assets/autocomplete.gif]
+
+
 ## 1.0.0
 
 - Initial release